feat(errors): extract errors and helpers into separate package
BREAKING CHANGE:
