+++ Makefile
@@ -9,7 +9,7 @@ compile-proto:
 
 	protoc --go_out=. --go_opt=paths=source_relative \
 		--go-grpc_out=. --go-grpc_opt=paths=source_relative \
-		./errors/errors.proto
+		./errors/proto/errors.proto
 
 	protoc --go_out=. --go_opt=paths=source_relative \
 		--go-grpc_out=. --go-grpc_opt=paths=source_relative \
