+++ context_helpers.go
@@ -4,13 +4,14 @@ import (
 	"context"
 
 	fctx "github.com/ri-nat/foundation/context"
+	ferr "github.com/ri-nat/foundation/errors"
 )
 
 // CheckAllScopesPresence checks if the context contains all specified scopes.
 // If any of the specified scopes is missing, it returns an "insufficient scope" error.
-func CheckAllScopesPresence(ctx context.Context, scopes ...string) *PermissionDeniedError {
+func CheckAllScopesPresence(ctx context.Context, scopes ...string) *ferr.PermissionDeniedError {
 	if !fctx.GetScopes(ctx).ContainsAll(scopes...) {
-		return NewInsufficientScopeAllError(scopes...)
+		return ferr.NewInsufficientScopeAllError(scopes...)
 	}
 
 	return nil
@@ -18,9 +19,9 @@ func CheckAllScopesPresence(ctx context.Context, scopes ...string) *PermissionDe
 
 // CheckAnyScopePresence checks if the context contains at least one of the specified scopes.
 // If none of the specified scopes are present, it returns an "insufficient scope" error.
-func CheckAnyScopePresence(ctx context.Context, scopes ...string) *PermissionDeniedError {
+func CheckAnyScopePresence(ctx context.Context, scopes ...string) *ferr.PermissionDeniedError {
 	if !fctx.GetScopes(ctx).ContainsAny(scopes...) {
-		return NewInsufficientScopeAnyError(scopes...)
+		return ferr.NewInsufficientScopeAnyError(scopes...)
 	}
 
 	return nil
