+++ grpc/errors_interceptor_test.go
@@ -1,51 +1,26 @@
-package foundation
+package grpc
 
 import (
 	"context"
-	"fmt"
-	"strings"
 	"testing"
 
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 
-	fctx "github.com/ri-nat/foundation/context"
+	ferr "github.com/ri-nat/foundation/errors"
 )
 
-func TestInternalError(t *testing.T) {
-	// Create a new internal error
-	err := NewInternalError(fmt.Errorf("test error"), "test")
-
-	// Check that the error message and details are correct
-	expectedSuff := "test: test error"
-	if !strings.HasSuffix(err.Error(), expectedSuff) {
-		t.Errorf("Expected error message to end with '%s', but got '%s'", expectedSuff, err.Error())
-	}
-
-	// Check that the error can be converted to a gRPC status error
-	status, ok := status.FromError(err)
-	if !ok {
-		t.Error("Expected a gRPC status error, but got a different error type")
-	} else if status.Code() != codes.Internal {
-		t.Errorf("Expected error code %s, but got %s", codes.Internal, status.Code())
-	} else if status.Message() != "internal error" {
-		t.Errorf("Expected error message '%s', but got '%s'", "internal error", status.Message())
-	}
-}
-
 func TestFoundationErrorToStatusInterceptor(t *testing.T) {
-	app := Init("test-app")
-
 	// Define a mock handler that returns an error
 	mockHandler := func(context.Context, interface{}) (interface{}, error) {
-		return nil, NewNotFoundError(nil, "test", "123")
+		return nil, ferr.NewNotFoundError(nil, "test", "123")
 	}
 
-	ctx := fctx.WithLogger(context.Background(), app.Logger)
+	ctx := context.Background()
 
 	// Call the interceptor with the mock handler and check that it returns the expected error status
-	_, err := app.foundationErrorToStatusUnaryInterceptor(ctx, nil, &grpc.UnaryServerInfo{}, mockHandler)
+	_, err := FoundationErrorToStatusUnaryInterceptor(ctx, nil, &grpc.UnaryServerInfo{}, mockHandler)
 	if err == nil {
 		t.Error("Expected an error, but got nil")
 	} else {
@@ -64,7 +39,7 @@ func TestFoundationErrorToStatusInterceptor(t *testing.T) {
 		return "test", nil
 	}
 
-	_, err = app.foundationErrorToStatusUnaryInterceptor(ctx, nil, &grpc.UnaryServerInfo{}, mockHandler)
+	_, err = FoundationErrorToStatusUnaryInterceptor(ctx, nil, &grpc.UnaryServerInfo{}, mockHandler)
 	if err != nil {
 		t.Errorf("Expected no error, but got %v", err)
 	}
@@ -74,7 +49,7 @@ func TestFoundationErrorToStatusInterceptor(t *testing.T) {
 		return nil, status.Error(codes.NotFound, "not found")
 	}
 
-	_, err = app.foundationErrorToStatusUnaryInterceptor(ctx, nil, &grpc.UnaryServerInfo{}, mockHandler)
+	_, err = FoundationErrorToStatusUnaryInterceptor(ctx, nil, &grpc.UnaryServerInfo{}, mockHandler)
 	if err == nil {
 		t.Error("Expected an error, but got nil")
 	} else {
